Another float to delete

+++ Research_code/research_code.py
@@ -9,7 +9,7 @@ import numpy as np
 def exp_decay(x:np.ndarray, # Input array that contains the indpendent variable.
     Amplitude: int|np.single, # Prefactor of the exponential decay.
     tau: int|np.single, # Decay constant.
-    x0: float|np.single, # Offset for the indpendent variable.
+    x0: int|np.single, # Offset for the indpendent variable.
     ) -> np.ndarray: # A new float array containing the exponentially decaying function.
 
     decayOutput = A * np.exp(-(x-x0)/tau1)
